requirements.txt
Flask
yt-dlp
openai-whisper
openai
reportlab
requests
python-dotenv